Grouped slice files together, updated store.ts and updated related type files/import addresses
