+++ package.json
@@ -42,6 +42,7 @@
     "@types/react-dom": "^18",
     "@types/uuid": "^9.0.8",
     "autoprefixer": "^10.0.1",
+    "cypress": "^13.9.0",
     "eslint": "^8",
     "eslint-config-next": "14.0.4",
     "jest": "^29.7.0",
