+++ types/form-types.ts
@@ -1,0 +1,7 @@
+import { Pet } from './pet-types';
+
+export interface FormPopupProps {
+  pets: Pet[];
+  formSelectedPet: Pet | null;
+  toggleFormPopup: (pet: Pet) => void;
+}
